// frontend/src/__generated__/models/MusicGameFacetSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * A game that has soundtrack tracks -- one entry of the album list.
 */
export type MusicGameFacetSchema = {
    rom_id: number;
    name: string;
    platform_id: number;
    platform_slug: string;
    platform_name: string;
    cover_url?: (string | null);
    count: number;
};