// frontend/src/__generated__/models/MusicPage_FacetValueSchema_.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FacetValueSchema } from './FacetValueSchema';
export type MusicPage_FacetValueSchema_ = {
    items: Array<FacetValueSchema>;
    total: number;
    limit: number;
    offset: number;
};