// frontend/src/__generated__/models/MusicPage_MusicGameFacetSchema_.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MusicGameFacetSchema } from './MusicGameFacetSchema';
export type MusicPage_MusicGameFacetSchema_ = {
    items: Array<MusicGameFacetSchema>;
    total: number;
    limit: number;
    offset: number;
};