// frontend/src/__generated__/models/MusicPage_MusicPlatformFacetSchema_.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MusicPlatformFacetSchema } from './MusicPlatformFacetSchema';
export type MusicPage_MusicPlatformFacetSchema_ = {
    items: Array<MusicPlatformFacetSchema>;
    total: number;
    limit: number;
    offset: number;
};