// frontend/src/__generated__/models/MusicPage_MusicTrackSchema_.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MusicTrackSchema } from './MusicTrackSchema';
export type MusicPage_MusicTrackSchema_ = {
    items: Array<MusicTrackSchema>;
    total: number;
    limit: number;
    offset: number;
};